Changed lua sql to use a thread whenever possible
All queries that are committed without providing a callback can be
executed in a different thread without disintegrating lua, but as
soon as the user passes a callback to `conn.Execute`, the sql api
has to flush the current threaded query queue, having the main thread
wait for it to finish, and then execute the given query right away in
the main thread itself. After it it completed we resume threaded execution.